# Conflicts:
#	src/GR.WebHosts/GR.Cms/GR.Cms.csproj

+++ src/GR.Extensions/GR.Ecommerce.Extension/GR.ECommerce.Razor/ViewModels/ProductVariationViewModel.cs
@@ -2,7 +2,6 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Text;
 
 namespace GR.ECommerce.Razor.ViewModels
 {
@@ -12,7 +11,7 @@ namespace GR.ECommerce.Razor.ViewModels
         public Guid? VariationId { get; set; }
         [Required]
         public Guid ProductId { get; set; }
-        public double Price { get; set; } = 0;
+        public decimal Price { get; set; } = 0;
         public IEnumerable<ProductVariationDetail> ProductVariationDetails { get; set; } = new List<ProductVariationDetail>();
     }
 }
